requirements.txt
livekit-agents
livekit-plugins-openai
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-noise-cancellation
mem0ai
duckduckgo-search
langchain_community
requests
python-dotenv
flask
fastapi
uvicorn[standard]
httpx
aiohttp
python-multipart
pydantic-settings
cryptography